refactor: reorganize components into common/ and photo/ subfolders
- Move UI primitives (CheckBox, IconButton, TextBox, ToggleSwitch) into components/common/
- Move photo-related components (DragSelectableGrid, EditImageSlider, ImageItem, ImageSlider) into components/photo/
- Move stack screens into app/(stack)/ group folder
- Update all import paths across app/ and components/
- Add diary generation console timer to loadingDiary.js
- Remove unused Variant21.png asset
