+++ app/(stack)/edit.js
@@ -12,12 +12,12 @@ import { useNavigation } from "@react-navigation/native";
 import { useLayoutEffect } from "react";
 import { useRouter } from "expo-router";
 import HeaderDate from "../../components/Header/HeaderDate";
-import IconButton from "../../components/IconButton";
-import TextBox from "../../components/TextBox";
+import IconButton from "../../components/common/IconButton";
+import TextBox from "../../components/common/TextBox";
 import characterList from "../../assets/characterList";
 import { usePhoto } from "../../contexts/PhotoContext";
 import CharacterPickerOverlay from "../../components/CharacterPickerOverlay";
-import EditImageSlider from "../../components/EditImageSlider";
+import EditImageSlider from "../../components/photo/EditImageSlider";
 import { useDiary } from "../../contexts/DiaryContext";
 import { useAuth } from "../../contexts/AuthContext";
 import Constants from "expo-constants";
