+++ app/(stack)/write.js
@@ -4,14 +4,14 @@ import { useRouter } from "expo-router";
 import { useNavigation } from "@react-navigation/native";
 import CharacterPickerOverlay from "../../components/CharacterPickerOverlay";
 import HeaderDate from "../../components/Header/HeaderDate";
-import IconButton from "../../components/IconButton";
-import TextBox from "../../components/TextBox";
+import IconButton from "../../components/common/IconButton";
+import TextBox from "../../components/common/TextBox";
 import characterList from "../../assets/characterList";
 import { useDiary } from "../../contexts/DiaryContext";
 import { usePhoto } from "../../contexts/PhotoContext";
 import { useAuth } from "../../contexts/AuthContext";
 import Constants from "expo-constants";
-import EditImageSlider from "../../components/EditImageSlider";
+import EditImageSlider from "../../components/photo/EditImageSlider";
 import { openGalleryAndAdd } from "../../utils/openGalleryAndAdd";
 import ConfirmModal from "../../components/Modal/ConfirmModal";
 import colors from "../../constants/colors";
