+++ app/(tabs)/profile/settings/waste.js
@@ -14,7 +14,7 @@ import { useRouter } from "expo-router";
 import Constants from "expo-constants";
 import { LinearGradient } from "expo-linear-gradient";
 import ConfirmModal from "../../../../components/Modal/ConfirmModal";
-import CheckBox from "../../../../components/CheckBox";
+import CheckBox from "../../../../components/common/CheckBox";
 import HeaderSettings from "../../../../components/Header/HeaderSettings";
 import { useAuth } from "../../../../contexts/AuthContext";
